Get Orders by table id

+++ backend/service/bill.go
@@ -2,18 +2,18 @@ package service
 
 type (
 	Bill struct {
-		ID        uint    `gorm:"primaryKey" json:"id"`
-		TableID   uint    `json:"table_id"`
+		ID        uint64  `gorm:"primaryKey" json:"id"`
+		TableID   uint64  `json:"table_id"`
 		Total     float64 `json:"total"`
 		CreatedAt int64   `json:"created_at"`
 	}
 
 	BillItem struct {
-		ID          uint     `gorm:"primaryKey" json:"id"`
+		ID          uint64   `gorm:"primaryKey" json:"id"`
 		ItemType    ItemType `json:"item_type"`
 		Description string   `json:"description"`
 		Price       float64  `json:"price"`
-		BillID      uint
+		BillID      uint64
 		Bill        Bill `json:"bill"`
 	}
 )
